Give Bedrock its own file, so a provider is a file
agent_providers.go held the generic seam AND all of Bedrock, which made the
answer to "how do I add a provider" into "append to the shared file". Now it is
"add a file plus one registry entry".

bedrock_setup.go takes the role ARN, the role-chaining session limit, the
region-to-geography prefix, the profile search, applyBedrockCreds, and both
providerSetup methods. agent_providers.go drops to 153 lines with NO AWS
imports at all — the measure that the seam is real.

Also completes the previous split, which left provider() and prepare() behind
in run_agent_step.go: the extraction stopped at the type declaration and did
not follow its methods.

Tests partition the same way, so adding a provider means adding a file and its
test file rather than editing shared ones.

One coupling stays on purpose: applyAgentModelEnv still checks
provider == AWSBedrock to write ANTHROPIC_MODEL instead of MODEL. That is not
a Bedrock fact but a claude-code one — claude-code reads the model from
ANTHROPIC_MODEL when its Bedrock switch is set — so it belongs beside
ApplyClaudeCodeUseBedrock in the shared catalogue, not here. Left for that
change rather than hidden behind a local abstraction.
